cleanup std flags
